@@ -1,50 +1,50 @@
 {
   "name": "ngry-chat",
-  "version": "0.0.0",
-  "license": "MIT",
+  "version": "0.0.1",
   "scripts": {
     "postinstall": "npm run build",
     "ng": "ng",
     "start": "ng serve",
-    "build": "ng build --prod",
+    "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "core-js": "^2.4.1",
-    "rxjs": "^5.5.6",
-    "twilio-chat": "^2.1.0",
-    "zone.js": "^0.8.19"
+    "@angular/animations": "~7.1.4",
+    "@angular/common": "~7.1.4",
+    "@angular/compiler": "~7.1.4",
+    "@angular/core": "~7.1.4",
+    "@angular/forms": "~7.1.4",
+    "@angular/platform-browser": "~7.1.4",
+    "@angular/platform-browser-dynamic": "~7.1.4",
+    "@angular/router": "~7.1.4",
+    "core-js": "^2.6.1",
+    "rxjs": "~6.3.3",
+    "tslib": "^1.9.3",
+    "twilio-chat": "~3.2.0",
+    "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "~1.7.0",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.8.0",
+    "@angular-devkit/build-angular": "~0.11.4",
+    "@angular/cli": "~7.1.4",
+    "@angular/compiler-cli": "~7.1.4",
+    "@angular/language-service": "~7.1.4",
+    "@types/jasmine": "~3.3.5",
+    "@types/jasminewd2": "~2.0.6",
+    "@types/node": "~10.12.18",
+    "codelyzer": "~4.5.0",
+    "jasmine-core": "~3.3.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "~3.1.4",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage-istanbul-reporter": "~2.0.4",
+    "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "protractor": "~5.4.2",
+    "ts-node": "~7.0.1",
+    "tslint": "~5.12.0",
+    "typescript": "~3.1.6"
   }
 }
