@@ -1,7 +1,7 @@
 import {Injectable} from '@angular/core';
 import {HttpClient, HttpHeaders} from '@angular/common/http';
 import {SentimentResult} from '../models/sentiment-result';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {LanguageCode} from '../models/language-code-enum';
 import {environment} from '../../environments/environment';
 import {SentimentRequest} from '../models/sentiment-request';
@@ -27,11 +27,7 @@ export class TextAnalysisService {
    * This method returns a json with documents being an array with the first (and only) element
    * being the result score for the given message. The important part here is the score, which says
    * if the message has a negative or positive mood. Below .5 means it's a negative tone if it is
-   * above .5 it is a more positive message.
-   *
-   * @param {string} msg
-   * @param {LanguageCode} language
-   * @returns {Observable<SentimentResult>}
+   * above .5 it is a more positive message.   *
    */
   getSentimentScoreForText(msg: string, language: LanguageCode): Observable<SentimentResult> {
     return this.httpClient
